../weather.js: (parseData): Move function into another file and import it using commonjs
../helper.js:
Add function parseData originally from weather.js file.
Export it using commonjs
