+++ scripts/helper.js
@@ -1,0 +1,15 @@
+function parsedData(data) {
+    let today = data;
+    let tomorrow = data.forecast.forecastday[1].day;
+    let dayAfter = data.forecast.forecastday[2].day;
+
+    return {
+        today: today,
+        tomorrow: tomorrow,
+        dayAfter: dayAfter,
+    };
+}
+
+module.exports = {
+    parsedData,
+};
